ppvm_runtime::traits

Trait ACMapConsume

pub trait ACMapConsume {
    // Required method
    fn consume(&mut self, dest: &mut Self);
}

Merge two maps with accumulation: drain dest into self, summing values that share a key.

Required Methods

fn consume(&mut self, dest: &mut Self)

Drain dest into self, accumulating values on key collision.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.